+++ spec/support/test_factories.rb
@@ -192,7 +192,9 @@ class TestFactories
       {
         name: "User-#{seq}",
         gender: 'm',
-        group_id: overrides[:group]&.id || overrides[:group_id] || factories.field_service_groups.random_or_create.id
+        group_id: overrides[:group]&.id ||
+          overrides[:group_id] ||
+          factories.field_service_groups.random_or_create.id
       }.merge(overrides)
     end
   end
